requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
drf-yasg==1.21.7
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-django==4.5.2
locust==2.16.1
django-debug-toolbar==4.2.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-modelcluster==6.3
django-taggit==5.0.1
djangorestframework-simplejwt